+++ tests/Jroc.Test262.Tests/built-ins/Array/prototype/find/ExecutionTests.cs
@@ -27,4 +27,11 @@ public class ExecutionTests : InMemoryExecutionTestsBase
     [Fact(DisplayName = "return-abrupt-from-property")]
     public Task return_abrupt_from_property()
         => ExecutionTestFromFile("return-abrupt-from-property");
+
+    [Fact(DisplayName = "resizable-buffer-grow-mid-iteration.js")]
+    public Task resizable_buffer_grow_mid_iteration() => ExecutionTestFromFile("resizable-buffer-grow-mid-iteration");
+    [Fact(DisplayName = "resizable-buffer-shrink-mid-iteration.js")]
+    public Task resizable_buffer_shrink_mid_iteration() => ExecutionTestFromFile("resizable-buffer-shrink-mid-iteration");
+    [Fact(DisplayName = "resizable-buffer.js")]
+    public Task resizable_buffer() => ExecutionTestFromFile("resizable-buffer");
 }