+++ tests/Jroc.Test262.Tests/built-ins/Array/prototype/toLocaleString/ExecutionTests.cs
@@ -1,0 +1,12 @@
+using Jroc.Test262.Tests.built_ins;
+
+namespace Jroc.Test262.Tests.built_ins.Array.prototype.toLocaleString;
+
+public class ExecutionTests : DiskExecutionTestsBase
+{
+    public ExecutionTests() : base("built_ins.built-ins.Array.prototype.toLocaleString") { }
+
+    [Fact(DisplayName = "resizable-buffer.js")]
+    public Task resizable_buffer() => ExecutionTestFromFile("resizable-buffer");
+
+}